Fixed battery calculations

+++ src/batteryState/batteryState.c
@@ -19,7 +19,7 @@ uint16_t batteryADCRead(uint8_t channel) {
 	return ADC;
 }
 
-int map(int x, int in_min, int in_max, int out_min, int out_max) {
+int map(float x, float in_min, float in_max, int out_min, int out_max) {
 	return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
 }
