+++ src/batteryState/batteryState.h
@@ -6,7 +6,7 @@ extern int mappedVoltage;
 
 void batteryADCInit();
 uint16_t batteryADCRead(uint8_t channel);
-int map(int x, int in_min, int in_max, int out_min, int out_max);
+int map(float x, float in_min, float in_max, int out_min, int out_max);
 void batteryADC();
 
 #endif
